% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{a_point}
\alias{a_point}
\title{Point}
\usage{
a_point(size, ...)
}
\arguments{
\item{size}{Point size.}

\item{...}{Any aframer element or option.}
}
\description{
Add a point.
}
\examples{
library(aframer)

browse_aframe(
  a_scene(
    a_dependency(),
    apoint_dependency(),
    a_point(5, color = "blue", perspective = FALSE, position = "0 1 -200"),
    a_point(5, color = "red", perspective = FALSE, position = "5 0 -250"),
    a_point(2, color = "black", perspective = FALSE, position = "-7 3 -100")
  )
)

}
