% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{a_stars}
\alias{a_stars}
\title{Add stars}
\usage{
a_stars(...)
}
\arguments{
\item{...}{Any aframer element or option.}
}
\description{
Add stars background.
}
\examples{
library(aframer)

browse_aframe(
  a_scene(
    a_dependency(),
    apoint_dependency(),
    a_sky(color = "#000"),
    a_stars()
  )
)

}
