% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{apoint_dependency}
\alias{apoint_dependency}
\title{Dependency}
\usage{
apoint_dependency(cdn = FALSE)
}
\arguments{
\item{cdn}{Whether to use local or CDN files.}
}
\description{
Add apoint dependency.
}
\note{
Source \link[aframer]{a_dependency} \emph{before} \code{apoint_dependency}.
}
\examples{
library(aframer)

browse_aframe(
  a_scene(
    a_dependency(),
    apoint_dependency()
  )
)

}
